Working on player controls ands ai behaviors
AI now roams in the rough area of the player and will switch from
seeking to roaming when the player gets too far away.

Player controllers are also mostly in, just need to figure out how to do
the rotation about the x to make it look like a real ship banking.